ci: GitHub Actions, including the parity gate that unit tests cannot replace
There was no CI at all, and scripts/check.sh claimed to "mirror what CI runs" against nothing.

ci.yml is the conventional Rust gate on a four-target matrix: fmt, clippy -D warnings, release
build and release tests on linux-x86_64, linux-aarch64, macos-arm64 and macos-x86_64. The x86_64
runners matter more than they look: this repo is developed on Apple Silicon, so without them the
AVX2 kernels ship untested. bwa-gpu is a macos-only dependency and is simply absent on Linux.

parity.yml is the one that earns its keep. `cargo test` would have caught NONE of the four
byte-identity divergences fixed in 7673bb9: they lived in the `-N` and `-w` option paths and past
2^23 read pairs. This job downloads the pinned bwa-mem2 oracle, generates deterministic reads and
compares SAM records byte for byte across all 49 harness cases. It is proven to have teeth: run
the same harness against a pre-fix binary on this exact fixture and it fails `-N 1` and `-w 1`.

Supporting changes:

- scripts/make_test_reads.py generates the reads instead of committing a binary fixture. It uses
  the drand48 LCG rather than Python's `random`, whose algorithm is not guaranteed stable across
  releases, so a CI failure reproduces locally from the same command line. Its docstring states
  the limit plainly: reads simulated from a reference are mostly unique-locus, which is exactly
  the blind spot that once let the missing discard pass score 5000/5000 while real reads diverged.
  This job is a regression net, not proof of byte-identity; proof is scripts/giab30x_pe.sh.
- opt_parity.sh takes IDX/R1/R2/M2 from the environment so CI can point it at testdata/tiny.
- Cargo.toml gains an explicit [workspace.lints.clippy] allow-list, one entry per lint the tree
  does not yet satisfy, each with the reason. This is what makes `-D warnings` enforceable today
  while still failing on any NEW lint. Burn the list down rather than extending it.
- The workspace `repository` pointed at a personal fork rather than IPNP-BIPN.

Verified before committing: fmt clean, clippy -D warnings clean, 71/71 tests, 49/49 parity, and
SAM output byte-identical to the pre-change binary on 1,006,597 records (the clippy autofix
touched 12 lines of real code, so that check was not a formality).

+++ crates/bwa-neon/src/lib.rs
@@ -450,7 +450,7 @@ mod tests {
                                 // Target offset to corrupt; the base there is bumped by 1..=3
                                 // modulo 4, which is guaranteed to change it to a different base.
                                 let p = (next() as usize) % t.len();
-                                t[p] = ((t[p] + 1 + (next() % 3) as u8) % 4) as u8;
+                                t[p] = (t[p] + 1 + (next() % 3) as u8) % 4;
                             }
                         }
                         t.extend((0..extra).map(|_| (next() % 4) as u8));
